backend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.2
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-graphql-jwt==0.3.4
django-js-asset==1.2.2
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.0.1
promise==2.3
pycodestyle==2.8.0
PyJWT==2.3.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
